--- a/keras/layers/core/dense.py
+++ b/keras/layers/core/dense.py
@@ -250,7 +250,9 @@ class Dense(Layer):
     def quantized_call(self, inputs):
         @ops.custom_gradient
         def matmul_with_inputs_gradient(inputs, kernel, kernel_scale):
-            def grad_fn(upstream):
+            def grad_fn(*args, upstream=None):
+                if upstream is None:
+                    (upstream,) = args
                 float_kernel = ops.divide(
                     ops.cast(kernel, dtype=self.compute_dtype),
                     kernel_scale,