--- a/keras/layers/core/dense.py
+++ b/keras/layers/core/dense.py
@@ -192,8 +192,8 @@ class Dense(Layer):
         # Do nothing if the layer isn't yet built
         if not self.built:
             return
-        # The keys of the `store` is determined because the defualt ordering
-        # will be changed after quantization
+        # The keys of the `store` will be saved as determined because the
+        # default ordering will change after quantization
         kernel_value, kernel_scale = self._get_kernel_with_merged_lora()
         store["0"] = kernel_value
         if self.use_bias:
@@ -207,8 +207,8 @@ class Dense(Layer):
         # Do nothing if the layer isn't yet built
         if not self.built:
             return
-        # The keys of the `store` is determined because the defualt ordering
-        # will be changed after quantization
+        # The keys of the `store` will be saved as determined because the
+        # default ordering will change after quantization
         self._kernel.assign(store["0"])
         if self.use_bias:
             self.bias.assign(store["1"])