--- a/keras/layers/core/einsum_dense.py
+++ b/keras/layers/core/einsum_dense.py
@@ -245,6 +245,11 @@ class EinsumDense(Layer):
         self.lora_rank = rank
 
     def save_own_variables(self, store):
+        # Do nothing if the layer isn't yet built
+        if not self.built:
+            return
+        # The keys of the `store` is determined because the defualt ordering
+        # will be changed after quantization
         kernel_value, kernel_scale = self._get_kernel_with_merged_lora()
         store["0"] = kernel_value
         if self.bias is not None:
@@ -253,6 +258,13 @@ class EinsumDense(Layer):
             store["2"] = kernel_scale
 
     def load_own_variables(self, store):
+        if not self.lora_enabled:
+            self._check_load_own_variables(store)
+        # Do nothing if the layer isn't yet built
+        if not self.built:
+            return
+        # The keys of the `store` is determined because the defualt ordering
+        # will be changed after quantization
         self._kernel.assign(store["0"])
         if self.bias is not None:
             self.bias.assign(store["1"])