--- a/keras/layers/core/einsum_dense.py
+++ b/keras/layers/core/einsum_dense.py
@@ -248,8 +248,8 @@ class EinsumDense(Layer):
         # Do nothing if the layer isn't yet built
         if not self.built:
             return
-        # The keys of the `store` is determined because the defualt ordering
-        # will be changed after quantization
+        # The keys of the `store` will be saved as determined because the
+        # default ordering will change after quantization
         kernel_value, kernel_scale = self._get_kernel_with_merged_lora()
         store["0"] = kernel_value
         if self.bias is not None:
@@ -263,8 +263,8 @@ class EinsumDense(Layer):
         # Do nothing if the layer isn't yet built
         if not self.built:
             return
-        # The keys of the `store` is determined because the defualt ordering
-        # will be changed after quantization
+        # The keys of the `store` will be saved as determined because the
+        # default ordering will change after quantization
         self._kernel.assign(store["0"])
         if self.bias is not None:
             self.bias.assign(store["1"])