… keras/layers/core/einsum_dense_test.py
… keras/layers/layer.py