@mattdangerw 2024-03-22
Is there a way we could try this out with an LLM? E.g. adapt this guide -> https://ai.google.dev/gemma/docs/lora_tuning

IIUC, this should probably allow us to fine tune a gemma 7b model checkpoint on less than 16gb GPU RAM, because lora will essentially zero out the size of optimizer variables relative to model weights, and quantizing our weight should bring us to a little over 8gb of space.

An end to end test with a massive model might validate a lot.