@mattdangerw 2024-03-22
How do we want to handle embeddings and quantization? Embeddings are usually the biggest individual memory hogs for models that might want quantization. We might want to add some quant support to our layer (though does not need to be this PR!).

https://github.com/google/gemma_pytorch/blob/cf8658c186255379194ba5b62612321eacde1b6b/gemma/model.py#L132-L154
https://pytorch.org/docs/stable/generated/torch.ao.nn.quantized.Embedding.html