@fchollet 2024-03-22
Since `Embedding` is just a lookup table, presumably just storing the weights in `int8` should do it? There's no input scaling either (inputs are integer indices).