@mattdangerw 2024-03-22
> Since Embedding is just a lookup table, presumably just storing the weights in int8 should do it? There's no input scaling either (inputs are integer indices).

Yeah I think it should be pretty simple.

Then there's the question of lora + quantization + an embedding layer. I don't think practically doing lora + quantization will be that important, as most people probably use lora with any embeddings frozen, but it might be worth adding for consistency (since we have enable_lora on the layer).