@james77777778 2024-03-23
> Is there a way we could try this out with an LLM? E.g. adapt this guide -> https://ai.google.dev/gemma/docs/lora_tuning

I have encountered an issues to try this PR with KerasNLP:
When I trying to load a quantized model using `keras.saving.load_model(...)`, I noticed that KerasNLP omits the `dtype` and always use the global dtype policy during loading.

This issue will cause the following to fail:
```python
preprocessor = keras_nlp.models.GPT2CausalLMPreprocessor.from_preset(
    "gpt2_base_en", sequence_length=128
)
lora_model = keras_nlp.models.GPT2CausalLM.from_preset(
    "gpt2_base_en", preprocessor=preprocessor
)
lora_model.quantize("int8")
lora_model.save("model_int8.keras")
reloaded_model = keras.saving.load_model("model_int8.keras")  # <- this line
```

The above is neccesary because tensorflow doesn't release GPU memory after using `quantize("int8")`. Therefore, to accurately record the peak GPU memory usage, I need to load the int8 model directly instead of initializing a float model and then quantizing it.

> How do we want to handle embeddings and quantization?

I can add  `quantize` to `Embedding` layer in another PR. It should be feasible.


Hi @mattdangerw , @fchollet 
Do we have a good method for measuring GPU memory usage in JAX? I fail to test this PR with KerasNLP models due to `dtype` issue when loading the model.

Thanks!