@james77777778 2024-03-24
I just realized that I need to manually call  `gc.collect()` to release GPU memory after the quantization.
I have added the result of finetuning GPT2. Please refer to https://github.com/keras-team/keras/pull/19356#issue-2201822208