@fchollet 2024-03-24
Thanks for the update and for the GPT-2 numbers!

> I just realized that I need to manually call gc.collect() to release GPU memory after the quantization.

Is this something we should do in the framework code, then?

Have you been able to try Gemma 2B?