@james77777778 2024-03-25
> Is this something we should do in the framework code, then?

This should be a harmless addition. I have updated the code and verified its effectiveness.

> Have you been able to try Gemma 2B?

Unfortunately, I failed to fit Gemma 2B with my rig. (12GB 4070...)
I have added an option `--model [gpt2|gemma]` to the script. Perhaps someone can try it out to obtain the result.

BTW, the training speed is improved after #19368 

|model|quantized|`compute_dtype`|training speed|peak memory usage|
|-|-|-|-|-|
|GPT2|❌|float32|120ms|4.728GB|
|GPT2|int8|float32|112ms|3.994GB|
|GPT2|❌|bfloat16|86ms|3.423GB|
|GPT2|int8|bfloat16|88ms|2.745GB|

Finetuning a quantized model might be faster in float32 and is competitive in bfloat16